support azure blob image upload
