+++ apps/sim/lib/copilot/api.ts
@@ -42,7 +42,7 @@ export interface CopilotChat {
  */
 export interface MessageFileAttachment {
   id: string
-  s3_key: string
+  key: string
   filename: string
   media_type: string
   size: number
